requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# AI Model APIs
openai==1.54.0
google-generativeai==0.8.3
groq==0.11.0

# Configuration
python-dotenv==1.0.0
jproperties==2.1.1

# Utilities
requests==2.31.0

# Fix for compatibility
httpx==0.27.0
